+++ version-gate-testkit/src/test/java/io/github/kbarseghyan/versiongate/testkit/InMemoryControlStoreContractTest.java
@@ -1,0 +1,12 @@
+package io.github.kbarseghyan.versiongate.testkit;
+
+import io.github.kbarseghyan.versiongate.port.ControlStore;
+import java.time.Clock;
+
+final class InMemoryControlStoreContractTest extends ControlStoreContract {
+
+  @Override
+  protected ControlStore createControlStore(Clock clock) {
+    return new InMemoryControlStore(clock);
+  }
+}